web/requirements.txt
firebase-admin==5.*
flask==2.*
google-cloud-secret-manager==2.*
gunicorn==20.*
slack-sdk==3.*
pyyaml==6.*
